post-debugging updates, tdd prep
