+++ 4_debugging/exercises/3_figure_it_out/tests/test_reverse_words.py
@@ -23,20 +23,22 @@ class TestReverseWords(unittest.TestCase):
     
     def test_empty_string(self):
         """It should return empty string for empty input"""
-        self.assertEqual(reverse_words(""), "")
+        self.assertEqual(reverse_words(""), "") # probably ok
     
     def test_one_word(self):
         """It should return the same string for one word"""
-        # BUG: Test doesn't account for extra space bug
-        self.assertEqual(reverse_words("hello"), "hello")
+        self.assertEqual(reverse_words("hello"), "hello") # probably ok
     
     def test_two_words(self):
         """It should reverse two words"""
-        # BUG: Test expects wrong output format
-        self.assertEqual(reverse_words("hello world"), "world,hello")
+        self.assertEqual(reverse_words("hello world"), "world hello") # probably ok
     
-    def test_multiple_spaces(self):
-        """It should handle multiple spaces"""
-        self.assertEqual(reverse_words("hello   world"), "world hello")
+    def test_two_spaces(self):
+        """It should handle two spaces"""
+        self.assertEqual(reverse_words("hello  world"), "world  hello") # probably ok
+
+    def test_three_spaces(self):
+        """It should handle three spaces"""
+        self.assertEqual(reverse_words("hello   world"), "world   hello") # probably ok
     
     # write more tests!
