+++ 5_tdd_with_llms/README.md
@@ -39,19 +39,25 @@ master all of the skills introduced in this workshop.
 
 </details>
 
-### TDD with LLMs
+### TDD without LLMs
 
-- 🥚 You can translate an open-ended problem into a complete docstring.
-- 🥚 You can use an LLM to generate code by providing a clear prompt and code
-  for context.
-- 🥚 You can comment, understand and review code generated by an LLM, judging if
-  it is correct or not.
+- 🥚 You can translate an open-ended problem into a complete docstring with doctests.
+- 🥚 You write unit tests for a function _before_ the function exists.
+- 🥚 You can stay focused on passing one test at a time when writing a function.
 - 🥚 You run your unit tests every time you make any change to your function,
   making sure they still pass.
 - 🥚 Every time you want to change a function's behavior, you first update the
   tests before changing the function.
 - 🐣 When a test fails, you can determine if it was the test or the function
   that is not correct.
+- 🐣 You can iteratively improve your function's implementation and still pass your test suite.
+
+### TDD with LLMs
+
+- 🥚 You can use an LLM to generate code by providing a clear prompt and code
+  for context.
+- 🥚 You can comment, understand and review code generated by an LLM, judging if
+  it is correct or not.
 - 🐣 You can collaborate with an LLM to find and fix bugs in your code.
 - 🐣 You can collaborate with an LLM to _refactor_ a program's strategy or
   implementation without modifying its behavior.
@@ -70,7 +76,7 @@ master all of the skills introduced in this workshop.
 - 🥚 You are aware that LLMs can, despite many guardrails, produce responses
   that perpetuate stereotypes and misconceptions.
 
-### Non-Objective: Prompt Engineering
+### Non-Objective: _Prompt Engineering_
 
 This workshop does not cover _prompt engineering_: how to carefully construct
 sentences and questions for the LLM to get better results. Prompt engineering is
