+++ 5_tdd_with_llms/examples/filter_and_sort_dictionaries/1b_manual_refactor_tests.py
@@ -1,17 +1,7 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on XX XX XX
-
-@author: Evan Cole + Chat GPT
-"""
-
-# ----- define function (normally in a separate file) -----
+import unittest
 
 
-def filter_and_sort_dictionaries(
-    list_to_sort: list[dict[str, str]], sort_key: str
-) -> list[dict[str, str]]:
+def f(lst, key):
     """Filters a list of dicts by key, and alphabetizes the entries by value.
 
     This function does not modify the argument list.
@@ -23,126 +13,110 @@ def filter_and_sort_dictionaries(
     Returns: a list of dicts where each dict contains the given key,
       and the dicts are sorted alphabetically by the value stored in the given key
 
-    >>> filter_and_sort_dictionaries([{'a':'z'},{'b':'y'},{'a':'x'}], 'a')
-    [{'a': 'x'}, {'a': 'z'}]
+    Raises: AssertionError
+        if the argument is not a list
+        if each item is not a dict
+        if any key in any dict is not a string
+        if any value in any dict is not a string
+
+    >>> f([{'a':'z'},{'b':'y'},{'a':'x'}], 'a')
+    [{'a':'x'},{'a','z'}]
 
 
-    >>> filter_and_sort_dictionaries([{'a':'z','b':'j'},{'b':'y'},{'b':'i','a':'x'}], 'b')
-    [{'b': 'i', 'a': 'x'}, {'a': 'z', 'b': 'j'}, {'b': 'y'}]
+    >>> f([{'a':'z','b':'j'},{'b':'y'},{'b':'i',a':'x'}], 'b')
+    [{'b':'i',a':'x'},{'a':'z','b':'j'},{'b':'y'}]
 
-    >>> filter_and_sort_dictionaries([{'a':'z'},{'b':'y'},{'a':'x'}], 'x')
-    []
     """
-    assert isinstance(list_to_sort, list), "Input must be a list"
-    for dictionary in list_to_sort:
-        assert isinstance(dictionary, dict), "List elements must be dictionaries"
-        for key, value in dictionary.items():
-            assert isinstance(key, str), "All keys must be strings"
-            assert isinstance(value, str), "All values must be strings"
+    assert isinstance(lst, list), "Input must be a list"
+    assert all(isinstance(d, dict) for d in lst), "List elements must be dictionaries"
+    # removed unnecessary assertion
+    assert all(
+        isinstance(d[key], str) for d in lst
+    ), f"Values for key '{key}' must be strings"
 
-    # create a new list with the filtered dicts, this will help avoid side effects
-    filtered_list = []
-    for dictionary in list_to_sort:
-        if sort_key in dictionary:
-            filtered_list.append(dictionary)
-
-    # sort the filtered dictionaries based on the value stored in the provided key
-    sorted_list = sorted(filtered_list, key=lambda dictionary: dictionary[sort_key])
-
-    return sorted_list
+    # implementation of the function goes here
+    pass
 
 
-# ----- test function (normally in a separate file) -----
-
-import unittest
-
-
-class TestFilterAndSortDictionaries(unittest.TestCase):
+class TestFilterAndSort(unittest.TestCase):
     """A test suite for filtering and sorting a list of dicts"""
 
     def test_emtpy_list(self):
         """Test case 1: Empty list"""
-        list_to_sort = []
-        result = filter_and_sort_dictionaries(list_to_sort, "a")
+        lst = []
+        result = f(lst, "a")
         expected = []
         self.assertEqual(result, expected)
 
     def test_list_with_one_dictionary(self):
         """Test case 2: List with one dictionary"""
-        list_to_sort = [{"a": "z"}]
-        result = filter_and_sort_dictionaries(list_to_sort, "a")
+        lst = [{"a": "z"}]
+        result = f(lst, "a")
         expected = [{"a": "z"}]
         self.assertEqual(result, expected)
 
     def test_many_dicts_same_key(self):
         """Test case 3: List with multiple dictionaries, all having the same key"""
-        list_to_sort = [{"a": "z"}, {"a": "y"}, {"a": "x"}]
-        result = filter_and_sort_dictionaries(list_to_sort, "a")
+        lst = [{"a": "z"}, {"a": "y"}, {"a": "x"}]
+        result = f(lst, "a")
         expected = [{"a": "x"}, {"a": "y"}, {"a": "z"}]
         self.assertEqual(result, expected)
 
     def test_many_dicts_some_shared_keys(self):
         """Test case 4: List with multiple dictionaries, some missing the key"""
-        list_to_sort = [{"a": "z"}, {"b": "y"}, {"a": "x"}]
-        result = filter_and_sort_dictionaries(list_to_sort, "a")
+        lst = [{"a": "z"}, {"b": "y"}, {"a": "x"}]
+        result = f(lst, "a")
         expected = [{"a": "x"}, {"a": "z"}]
         self.assertEqual(result, expected)
 
     def test_many_dicts_no_shared_keys(self):
         """Test case 5: List with dictionaries having different keys"""
-        list_to_sort = [{"a": "z"}, {"b": "y"}, {"c": "x"}]
-        result = filter_and_sort_dictionaries(list_to_sort, "a")
+        lst = [{"a": "z"}, {"b": "y"}, {"c": "x"}]
+        result = f(lst, "a")
         expected = [{"a": "z"}]
         self.assertEqual(result, expected)
 
     def test_dicts_with_non_string_keys(self):
         """Test case 6: List with dictionaries having non-string values for the key"""
-        list_to_sort = [{"a": 1}, {"a": True}, {"a": "x"}]
+        lst = [{"a": 1}, {"a": True}, {"a": "x"}]
         with self.assertRaises(AssertionError):
-            filter_and_sort_dictionaries(list_to_sort, "a")
+            f(lst, "a")
 
     def test_dicts_with_other_non_string_keys(self):
         """Test case 7: List of dictionaries with non-string values for other keys"""
-        list_to_sort = [{"a": "z"}, {"b": 1}, {"c": True}]
+        lst = [{"a": "z"}, {"b": 1}, {"c": True}]
         with self.assertRaises(AssertionError):
-            filter_and_sort_dictionaries(list_to_sort, "a")
+            f(lst, "a")
 
     def test_list_with_non_dict_elements(self):
         """Test case 8: List with dictionaries having non-dict elements"""
-        list_to_sort = [{"a": "z"}, "b", {"a": "x"}]
+        lst = [{"a": "z"}, "b", {"a": "x"}]
         with self.assertRaises(AssertionError):
-            filter_and_sort_dictionaries(list_to_sort, "a")
+            f(lst, "a")
 
     def test_dicts_with_all_different_keys(self):
         """Test case 9: List with dictionaries having different keys and values"""
-        list_to_sort = [
-            {"a": "z"},
-            {"b": "y"},
-            {"c": "x"},
-            {"a": "w"},
-            {"b": "v"},
-            {"c": "u"},
-        ]
-        result = filter_and_sort_dictionaries(list_to_sort, "b")
+        lst = [{"a": "z"}, {"b": "y"}, {"c": "x"}, {"a": "w"}, {"b": "v"}, {"c": "u"}]
+        result = f(lst, "b")
         expected = [{"b": "v"}, {"b": "y"}]
         self.assertEqual(result, expected)
 
     def test_dicts_with_same_keys(self):
         """Test case 10: List with dictionaries having duplicate values for the key"""
-        list_to_sort = [{"a": "z"}, {"a": "y"}, {"a": "y"}, {"a": "x"}]
-        result = filter_and_sort_dictionaries(list_to_sort, "a")
+        lst = [{"a": "z"}, {"a": "y"}, {"a": "y"}, {"a": "x"}]
+        result = f(lst, "a")
         expected = [{"a": "x"}, {"a": "y"}, {"a": "y"}, {"a": "z"}]
         self.assertEqual(result, expected)
 
     def test_dicts_with_many_keys(self):
         """Test case 11: List with dictionaries having duplicate values for the key"""
-        list_to_sort = [
+        lst = [
             {"a": "z", "b": "hi"},
             {"a": "y", "c": "bonjour"},
             {"a": "y"},
             {"a": "x", "b": "hallo"},
         ]
-        result = filter_and_sort_dictionaries(list_to_sort, "b")
+        result = f(lst, "b")
         expected = [
             {"a": "x", "b": "hallo"},
             {"a": "z", "b": "hi"},
@@ -151,13 +125,13 @@ class TestFilterAndSortDictionaries(unittest.TestCase):
 
     def test_dicts_with_many_long_keys(self):
         """Test case 12: List with dictionaries having duplicate values for the key"""
-        list_to_sort = [
+        lst = [
             {"alphabet": "z", "bodega": "hi"},
             {"alphabet": "y", "Calliope": "bonjour"},
             {"alphabet": "y", "Calliope": "goede dag"},
             {"alphabet": "x", "bodega": "hallo"},
         ]
-        result = filter_and_sort_dictionaries(list_to_sort, "Calliope")
+        result = f(lst, "Calliope")
         expected = [
             {"alphabet": "y", "Calliope": "bonjour"},
             {"alphabet": "y", "Calliope": "goede dag"},
@@ -166,21 +140,23 @@ class TestFilterAndSortDictionaries(unittest.TestCase):
 
     def test_empty_dictionaries(self):
         """Test case 13: A list with some empty dictionaries"""
-        list_to_sort = [
+        lst = [
             {"a": "z"},
             {},
             {"a": "y"},
-            {"a": "x"},
+            {
+                "a": "x",
+            },
             {},
         ]
-        result = filter_and_sort_dictionaries(list_to_sort, "a")
+        result = f(lst, "a")
         expected = [{"a": "x"}, {"a": "y"}, {"a": "z"}]
         self.assertEqual(result, expected)
 
     def test_does_not_modify_argument_list(self):
         """Test case 14: Function does not modify the list argument"""
         lst = [{"a": "x"}, {"a": "y"}, {"a": "z"}]
-        filter_and_sort_dictionaries(lst, "a")
+        f(lst, "a")
         self.assertEqual(lst, [{"a": "x"}, {"a": "y"}, {"a": "z"}])
