+++ 5_tdd_with_llms/examples/filter_and_sort_dictionaries/2a_generated_function.py
