+++ 5_tdd_with_llms/examples/filter_and_sort_dictionaries/2c_generated_refactor_function.py
@@ -1,6 +1,7 @@
 import unittest
 
 
+# I asked the AI for a more explicit, imperative strategy
 def f(lst, key):
     """Filters a list of dicts by key, and alphabetizes the entries by value.
 
@@ -13,6 +14,12 @@ def f(lst, key):
     Returns: a list of dicts where each dict contains the given key,
       and the dicts are sorted alphabetically by the value stored in the given key
 
+    Raises: AssertionError
+        if the argument is not a list
+        if each item is not a dict
+        if any key in any dict is not a string
+        if any value in any dict is not a string
+
     >>> f([{'a':'z'},{'b':'y'},{'a':'x'}], 'a')
     [{'a':'x'},{'a','z'}]
 
@@ -23,16 +30,19 @@ def f(lst, key):
     """
     assert isinstance(lst, list), "Input must be a list"
     assert all(isinstance(d, dict) for d in lst), "List elements must be dictionaries"
-    # removed extra assertion
-    # fixed error in assertion, asked AI to check if all keys and values are strings
-    #   manually refactored the loop to assert
     for dictionary in lst:
         for key, value in dictionary.items():
             assert isinstance(key, str), "some keys are not strings"
             assert isinstance(value, str), "some values are not strings"
 
-    filtered_list = [d for d in lst if isinstance(d, dict) and key in d]
-    sorted_list = sorted(filtered_list, key=lambda x: x[key])
+    filtered_list = []
+    for d in lst:
+        if key in d:
+            filtered_list.append(d)
+
+    sorted_list = []
+    for d in sorted(filtered_list, key=lambda x: x[key]):
+        sorted_list.append(d)
 
     return sorted_list
