--- 5_tdd_with_llms/examples/separated_even_odd/README.md
@@ -1,6 +1,0 @@
-# Separate Even and Odds
-
-> This example was completed as a full-class exercise in the workshop
-
-A function that takes a list of numbers and returns two lists, one all even the
-other all odd.
