+++ 5_tdd_with_llms/exercises/is_in.py
@@ -1,0 +1,6 @@
+""" Is in
+
+Write a function that takes in a string and two lists of strings. 
+It will return true if the item is in _at least one_ of the lists.
+
+"""
