+++ 5_tdd_with_llms/exercises/is_in_both.py
@@ -1,0 +1,6 @@
+""" Is in Both
+
+Write a function that takes in a string and two lists of strings. 
+It will return true if the item is in _both_ of the lists.
+
+"""
