+++ 5_tdd_with_llms/exercises/is_in_one.py
@@ -1,10 +1,6 @@
-""" Exclusive Or
+""" Is in one
 
 Write a function that takes in a string and two lists of strings. 
 It will return true if the item is in _only one_ of the lists.
 
 """
-
-
-def exclusive_or():
-    pass
