+++ 5_tdd_with_llms/exercises/repeat_character.py
@@ -4,7 +4,3 @@ Write a function that takes in a string, a single character, and a number.
 The function returns a string with each occurrence of the character repeated n times.
 
 """
-
-
-def repeat_character():
-    pass
