+++ 5_tdd_with_llms/exercises/sum_evens_and_odds.py
@@ -4,7 +4,3 @@ Write a function that takes a list of numbers
 and returns a dictionary with sums of the even and odd numbers in the list.
 
 """
-
-
-def sum_evens_and_odds():
-    pass
