+++ 5_tdd_with_llms/lesson_plan.md
@@ -1,32 +1,195 @@
-# Workshop Outline
+class: middle, center
 
-## Workshop Overview _(all together)_
+<!-- this file is written for remark: https://github.com/gnab/remark/wiki -->
 
-The workshop instructor will introduce the main concepts of this workshop:
+# Test-Driven Development with Large Language Models
 
-- Introduce the workshop's [learning objectives](./README.md#learning-objectives).
-- Discuss some trade-offs to consider when studying and developing with an LLM.
-- Talk through the [CoAIthoring Workflow](./coaithoring-workflow.md).
-- Demonstrate how to do TDD with an LLM by developing a function that adds two
-  numbers
-  - We will use [phind](https://phind.com) because you do not need to create an
-    account.
-- Introduce the group exercises.
+## _TDD with LLMs_
 
-## Practice CoAIthoring _(small groups)_
+<br />
 
-Each small group will choose one of the [/exercises](./exercises/) and develop
-tests + a solution using [phind](https://phind.com) for support.
+<img alt="Emerging Talent Logo" src="../.assets/emerging_talent_logo.png" height="50%"  width="50%">
 
-You can start with any exercise you like, and move on to another if you finish.
+---
 
-## Discussion _(all together)_
+class: middle, center
 
-Back together, you will have an informal discussion with the other groups and
-the workshop leader.
+## Knowledge Sharing: _Slido_
 
-1. Each group will have 2-3 minutes to share:
-   - One thing they couldn't figure out
-   - One surprising thing they learned
-   - One thing they'd like to discuss with the full class
-2. Discuss!
+---
+
+class: middle
+
+## Agenda
+
+- **Learning Objectives**
+
+- **Zone of Proximal Development (_ZPD_)**
+
+- **Test-Driven Development (_TDD_)**
+
+- **Large Language Models (_LLMs_)**
+
+- **TDD with LLMs**
+
+- **The Workflow**
+
+- **Breakout Groups: _./exercises_**
+
+- **Discussion + Q&A**
+
+---
+
+class: middle
+
+## Learning Objectives
+
+- You can use the ZPD model to choose which challenges to study
+
+- You understand the value of writing tests _before_ writing code
+
+- You can explain the TDD cycle and are ready to practice it on your own
+
+- You understand that LLMs do not replace good development habits
+
+- You understand how to integrate LLMs in your workflows
+  - Not as _replacements_ for good workflows
+
+- You can tell when a coding challenge is distractingly difficult for you right now
+  - Focus on your _development workflows_, not solving hard problems!
+
+- For self-study: _[objectives in the README](./README.md#learning-objectives)_
+
+---
+
+class: middle, center
+
+## Zone of Proximal Development (_ZPD_)
+
+[![zone of proximal development](../.assets/zone_of_proximal_development.png)](https://mavink.com/explore/Zone-of-Proximal-Development-Infographic)
+
+---
+
+class: middle
+
+## ZPD: _study advice_
+
+<a href="https://mavink.com/explore/Zone-of-Proximal-Development-Infographic" target="_blanks"><img alt="zone of proximal development" src="../.assets/zone_of_proximal_development.png" height="50%"  width="50%"></a>
+
+- You have of two separate ZPDs as you learn to program
+  - The **difficulty of challenges** you solve
+  - The **development process** you use to solve challenges
+
+- Focus on your  **development process**, not hard challenges
+  - Repeat the _entire process_ with _good practices_ and easy challenges
+  - Not fewer times with harder challenges with bad practices
+
+- Focus on coding challenges in your _can do now_ zone
+
+---
+
+class: middle, center
+
+## Test-Driven Development (_TDD_)
+
+![TDD cycle](./.assets/tdd_cycle.png)
+
+---
+
+class: middle
+
+## TDD: _write tests first?!_
+
+- Be sure you understand the function's _behavior_
+
+- Be confident your function behaves correctly
+
+- Unit tests are easier to write than functions!
+
+- Pace your development -> move slowly with confidence
+  - Spend time up front, save time later
+
+- You probably won't write tests _after_ your function works
+
+---
+
+class: middle
+
+## TDD: _incremental development_
+
+- Break the whole problem into smaller, simpler _behaviors_
+
+- Solve one small _behavior_ at a time
+
+- Improve your _strategy_ and _implementation_ before adding new _behaviors_
+
+---
+
+class: middle
+
+## Large Language Models (_LLMs_)
+
+_This workshop is not about how LLMs work, how to develop them, or their risks._
+
+<br />
+
+- You _are not_ required to use them for Foundations
+  - but we know most of you probably will
+
+- Think of them as _collaborators_, not _replacements_
+
+- LLMs can generate code, but they cannot understand the code for you
+
+- LLMs can make mistakes, it's your responsibility to catch them
+
+- LLMs are not free!
+  - Many require accounts, full access requires paid accounts
+  - They use _a lot_ of electricity and resources
+
+---
+
+class: middle
+
+## TDD with LLMs
+
+_TDD does not change! LLMs just help with some steps:_
+
+- You still need to understand the behavior
+
+- You still need to check and run your tests
+
+- You still need to step through and understand the function
+
+- You still need to decide if the strategy is overly complicated
+
+- You still need to decide if the implementation is clear
+
+- You still need to do code review with a checklist
+
+- The LLM only speeds up the bits where you write code
+
+---
+
+class: middle, center
+
+## ~~Breakout Groups: `./exercises`~~
+
+## Live Demo: _TDD without LLMs_
+
+---
+
+class: middle, center
+
+## Discussion + Q & A
+
+---
+
+class: middle, center
+
+# Thank You
+
+<br />
+
+<img alt="Emerging Talent Logo" src="../.assets/emerging_talent_logo.png" height="50%"  width="50%">
+
+---
