+++ 5_tdd_with_llms/workflow_with_llm.md
@@ -1,4 +1,4 @@
-# CoAIthoring Workflow
+# Workflow with LLM
 
 A suggested workflow to practice when developing a function with LLMs. With
 experience you can adapt this workflow to match your style:
@@ -6,19 +6,14 @@ experience you can adapt this workflow to match your style:
 0. Clearly describe the _behavior_ you want your function to have. Include a
    couple input/output examples.
    - Write your description as a _docstring_
-1. Ask the AI to generate a suite of black box unit tests for this docstring,
-   covering as many boundary cases as possible.
-   - You may need to reformat the tests to have one assertion per method.
-   - Read the tests to be sure you understand them and they are correct, make
-     changes or add test cases as necessary.
-   - You can also ask the AI to generate tests for defensive assertions.
-2. Ask the AI to generate a function that will pass your unit tests. Provide
-   your unit tests in your prompt as context.
-   - Test the function to make sure it passes. AI isn't always right!
-   - Debug any failing test cases. You can do this manually, and with help from
-     the LLM.
-   - If you still can't debug the function, explain to the AI which test cases
-     failed then ask it to regenerate the function.
+1. Write a suite of unit tests for this docstring covering as many boundary cases as possible.
+   - Double-check the tests to be sure you understand them and they are correct
+   - Include tests for defensive assertions.
+   - Run your tests to make sure they have no errors.
+2. Pick one unit test s and write the simplest possible function that will pass it.
+   - Test the function to make sure it passes.
+   - If you can't pass this test, focus on another one instead.  You can always come back to the one you skipped.s
+   - Step through the function to make sure it works how you think it does.
    - Repeat until all the tests pass.
 3. Study the passing code to understand it's _strategy_. Code is not
    maintainable if you don't understand it.
@@ -38,4 +33,4 @@ experience you can adapt this workflow to match your style:
      nested structures?
    - Don't forget to run your tests after every change you make!
 5. Give the code a final review using
-   [the checklist from Documenting and Testing](https://github.com/MIT-Emerging-Talent/documenting-and-testing/blob/main/code-review-checklists.md)
+   [the checklist from Documenting and Testing](../3_documenting_and_testing/code_review_checklist.md)
