+++ 6_recursion/3_use_recursion/2_write/tests/test_check_palindrome.py
@@ -8,16 +8,16 @@ from ..check_palindrome import check_palindrome
 
 class TestCheckPalindrome(unittest.TestCase):
     def test_empty_string(self):
-        self.assertTrue(check_palindrome(""))
+        self.assertTrue(check_palindrome(""), _)
 
     def test_single_character(self):
-        self.assertTrue(check_palindrome("a"))
+        self.assertTrue(check_palindrome("a"), _)
 
     def test_palindrome_even_length(self):
-        self.assertTrue(check_palindrome("abba"))
+        self.assertTrue(check_palindrome("abba"), _)
 
     def test_palindrome_odd_length(self):
-        self.assertTrue(check_palindrome("madam"))
+        self.assertTrue(check_palindrome("madam"), _)
 
     def test_not_a_palindrome(self):
-        self.assertFalse(check_palindrome("hello"))
+        self.assertFalse(check_palindrome("hello"), _)
